+++ package.json
@@ -63,17 +63,13 @@
     "@babel/preset-env": "7.1.0",
     "@babel/preset-flow": "7.0.0",
     "@babel/preset-react": "7.0.0",
-    "@emotion/babel-plugin-core": "0.6.9",
-    "@emotion/core": "10.0.0-really-unsafe-please-do-not-use.2",
-    "@emotion/css": "10.0.0-really-unsafe-please-do-not-use.2",
-    "@emotion/global": "0.7.0",
-    "@emotion/jsx": "0.7.0",
-    "@emotion/keyframes": "0.9.1",
-    "@emotion/style": "0.8.0",
+    "@emotion/core": "10.0.0-beta.6",
+    "@emotion/styled": "10.0.0-beta.6",
     "@reach/router": "1.3.0-alpha.1",
     "async-sema": "2.1.4",
     "autoprefixer": "9.3.1",
     "babel-loader": "8.0.4",
+    "babel-plugin-emotion": "10.0.0-beta.6",
     "babel-plugin-transform-react-remove-prop-types": "0.4.19",
     "benchmark": "2.1.4",
     "bootstrap": "4.1.3",
@@ -90,7 +86,7 @@
     "fg-loadjs": "1.0.0",
     "focus-trap": "3.0.0",
     "gzip-size": "5.0.0",
-    "immer": "1.7.3",
+    "immer": "1.7.4",
     "jszip": "3.1.5",
     "lodash-es": "4.17.11",
     "memoize-state": "2.0.4",
@@ -109,7 +105,7 @@
     "scheduler": "0.11.0-alpha.0",
     "scroll-into-view-if-needed": "2.2.20",
     "style-loader": "0.23.1",
-    "terser": "3.10.3",
+    "terser": "3.10.4",
     "thread-loader": "1.2.0",
     "three": "0.97.0",
     "typescript-styled-plugin": "0.12.0",
